+++ deltafi-ui/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Updated Server Sent Events endpoint (#312)
+- Added refresh button to Dashboard (#313)
 - Adjusted metadata fields that changed from KeyValues to maps
 
 ## [0.99.8]